+++ package.xml
@@ -2,8 +2,8 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>mission_control_package</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
+  <version>0.1.0</version>
+  <description>Package for the Mission Control node</description>
   <maintainer email="phg6386@thi.de">philipp</maintainer>
   <license>MIT</license>
